+++ WorldOfCells/objects/Monolith.java
@@ -10,7 +10,7 @@ import worlds.World;
 
 public class Monolith extends UniqueObject{
 	
-	public Monolith ( int __x , int __y , World __world )
+	public Monolith ( float __x , float __y , World __world )
 	{
 		super(__x,__y,__world);
 	}
@@ -22,9 +22,9 @@ public class Monolith extends UniqueObject{
         
         //gl.glColor3f(0.f+(float)(0.5*Math.random()),0.f+(float)(0.5*Math.random()),0.f+(float)(0.5*Math.random()));
         
-    	int x2 = (x-(offsetCA_x%myWorld.getWidth()));
+    	float x2 = (x-(offsetCA_x%myWorld.getWidth()));
     	if ( x2 < 0) x2+=myWorld.getWidth();
-    	int y2 = (y-(offsetCA_y%myWorld.getHeight()));
+    	float y2 = (y-(offsetCA_y%myWorld.getHeight()));
     	if ( y2 < 0) y2+=myWorld.getHeight();
 
     	float zoffset = myWorld.getLandscape().getZOffset();
