+++ WorldOfCells/objects/UniqueObject.java
@@ -10,19 +10,19 @@ import worlds.World;
 
 abstract public class UniqueObject // UniqueObject are object defined with particular, unique, properties (ex.: particular location)
 {
-	protected int x,y;
+	protected float x,y;
 	protected World world;
 	
-	public UniqueObject(int __x, int __y, World __world) 
+	public UniqueObject(float __x, float __y, World __world)
 	{
 		this.world = __world;
 		this.x = __x;
 		this.y = __y;
 	}
 	
-	public int[] getCoordinate()
+	public float[] getCoordinate()
 	{
-		int coordinate[] = new int[2];
+		float coordinate[] = new float[2];
 		coordinate[0] = this.x;
 		coordinate[1] = this.y;
 		return coordinate;
